LayoutTests/imported/w3c/web-platform-tests/navigation-timing/nav2_test_document_open-expected.txt

CONSOLE MESSAGE: line 60: TypeError: undefined is not an object (evaluating 'timing[navTiming2Attributes[i]]')
Description

This test validates window.performance.getEntriesByType("navigation") remains constant when a document is replaced using document.open.

This page should be loaded with a yellow frame below. It then replaces the document in that frame with a green document.

The test passes if all of the checks to performance.getEntriesByType("navigation") are correct and the frame below ends with a green background.



Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'timing[navTiming2Attributes[i]]')
